Adding more benchmarks
